Replace arparse arguement parser with Docopt
The arguement parser used was Argparse,
this is replaced with Docopt arguement parser.

Docopt:http://docopt.org

Docopt(it uses only re and sys)  is  a beautiful pythonic arguement parser. It allows Docstrings written as help
message to use the tool to be used to parse arguements.

Howdoi is starting to extend and include more options,so this will be easier
on the eye and brain.

Remove get_parser() and replace with parse_args(args=None)

Update Tests.