Display links using markdown syntax
Some answers on the StackExchange network include hyperlinks
to other content on the Internet. Ideally, those hyperlinks would serve
as "supplemental" reading, but there has been instances where the hyperlink
contains most of the answer (the so-called "link-only answers").

howdoi currently ignores all hyperlinks within the answer of the
question, preferring instead to display a raw-text version of the answer.
While this is suitable for most instances, there are a few cases
where you may want to dig deeper and follow those hyperlinks, especially
if you have encountered a "link-only answer". Rather than viewing the
StackExchange answer through the browser and risk distraction,
it would be preferable to see the hyperlinks directly in the terminal.