Fix bug where copy in links is not correctly gathered
The code previously assumed that the copy of the hyperlink is within
the 'a' tag, but it is possible that the copy may be nested within
another div, like a ```code``` tag. Therefore, we would return None
for the copy, even if there was some copy avaliable to display
to the user.

We fix this issue by wrapping the whole hyperlink as a "pquery object"
and then extracting the text from that object.

I also did some additional refactoring of the code as well,
since we no longer need to initialize a "pquery object" after
gathering the copy and href...since we need to initialize a
"pquery object" anyway to gather the copy.