A few readability suggestions
- Add a space after variable and before typescript types (`foo:HowdoiResult` -> `foo:
HowdoiResult`)
- Remove all semicolons (It's weird, I know, but they aren't necessary
in Typescript. I also changed the `.eslintrc` file to show errors if
semicolons are included)
- Prefer using `'` everywhere rather than `"` (added eslint rule)
- Make sure that there is always a space after a comma (added eslint
rule)

Some of these changes are personal preference, but by attaching an
associated eslint rule and requiring that linting pass before a PR can
be accepted, you can keep the codebase readable and uniform.