make error checking more robust in `_sanity_check()`, I was getting this error message:
```
b'Unable to reach google. Do you need to use a proxy?\n'
```

but since `_sanity_check()` was checking only for another error message, the check was
giving "Ok" result when actually I was offline.

Now instead of checking for error message which can change anyway and it is error prune,
check if we get replies using `-j` option.