Disable colorization check in Github
I tried a lot of different things to make it work

```
      env:
        PY_COLORS: '1'
        ANSIBLE_FORCE_COLOR: '1'
        FORCE_COLOR: '1'
```

I also tried

```
export TERM=xterm-256color;
```

But Github refuses to colorize. Seems to be a problem with Github
Actions not being a `tty`.

https://github.com/actions/runner/issues/241

This doesn't seem to be a priority for Github, so we simply disable
the test in that environment.